requirements.txt
openai
requests
Flask>=2.3.0
langchain-openai